requirements.txt
pandas
meteomatics
python-dotenv